test/web-platform-tests/to-upstream/cssom-view/Element-scrollLeft.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Element.prototype.scrollLeft</title>
<link rel=help href="https://drafts.csswg.org/cssom-view/#dom-element-scrollleft">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const el = document.createElement("p");
  assert_equals(el.scrollLeft, 0);

}, "scrollLeft should return 0 for an element disconnected from a document");
</script>
